feat: prevent the user from being able to delete themselves

+++ accessflow/views/admin/users/delete.py
@@ -1,6 +1,6 @@
 from flask import request, flash, redirect, url_for, abort
 from flask.views import View
-from flask_login import login_required
+from flask_login import login_required, current_user
 from accessflow.models.user_permission import UserPermission
 from accessflow.models.user import User
 from accessflow import db
@@ -10,9 +10,16 @@ class UserDeleteView(View):
     decorators = [login_required]
 
     def dispatch_request(self, user_id):
+        # The user shouldn't be able to delete themselves.
+        # The UI should prevent this from happening, but return a 500 error just in case
+        if current_user.id == user_id:
+            abort(500)
+
+        # Delete all records across all tables associated with the specified user
         user_permissions = UserPermission.query.filter(UserPermission.user_id == user_id).delete()
         user = User.query.filter(User.id == user_id).delete()
 
+        # Check the user ID is actually valid
         if not user_permissions or not user:
             abort(404)
